Moved test_make_mpl_fig() to test_plotting.py
